+++ mirai-core/src/commonMain/kotlin/net.mamoe.mirai/Bot.kt
@@ -291,10 +291,12 @@ abstract class Bot internal constructor(
     @Deprecated(
         "use member function.",
         replaceWith = ReplaceWith("event.reject(blackList)"),
-        level = DeprecationLevel.ERROR
+        level = DeprecationLevel.HIDDEN
     )
-    @JvmSynthetic
     abstract suspend fun rejectMemberJoinRequest(event: MemberJoinRequestEvent, blackList: Boolean = false)
+
+    @JvmSynthetic
+    abstract suspend fun rejectMemberJoinRequest(event: MemberJoinRequestEvent, blackList: Boolean = false, message: String = "")
 
     /**
      * 忽略加群验证（需管理员权限）